[Asynchronous] Move into separate directory
Three changes:

1. Move files related to `XCTestCase+Asynchronous` into a separate directory.
2. Move methods on `XCTestCase` that vend expectations for `Predicate` and
   `Notification` into their own files.
3. Modify the build script in order to traverse deeply nested files.

Besides general project hygiene, the motivation for this change is,
unintuitively, related to libdispatch: `import Dispatch` within
`XCTestCase+Asynchronous` would otherwise result in an error indicating
that `NSObjectProtocol` (used within the `expectation(forNotification:...)`
method) is an ambiguous reference. By keeping the source files small
and limiting the scope in which `Dispatch` is imported, we can sidestep
the error.
