+++ Sources/XCTest/Public/Asynchronous/XCTestCase+Asynchronous.swift
@@ -151,94 +151,6 @@ public extension XCTestCase {
             completionHandler(error)
         }
     }
-
-    /// Creates and returns an expectation for a notification.
-    ///
-    /// - Parameter notificationName: The name of the notification the
-    ///   expectation observes.
-    /// - Parameter object: The object whose notifications the expectation will
-    ///   receive; that is, only notifications with this object are observed by
-    ///   the test case. If you pass nil, the expectation doesn't use
-    ///   a notification's object to decide whether it is fulfilled.
-    /// - Parameter handler: If provided, the handler will be invoked when the
-    ///   notification is observed. It will not be invoked on timeout. Use the
-    ///   handler to further investigate if the notification fulfills the
-    ///   expectation.
-    @discardableResult func expectation(forNotification notificationName: String, object objectToObserve: AnyObject?, handler: XCNotificationExpectationHandler? = nil) -> XCTestExpectation {
-        let objectDescription = objectToObserve == nil ? "any object" : "\(objectToObserve!)"
-        let expectation = self.expectation(description: "Expect notification '\(notificationName)' from " + objectDescription)
-        // Start observing the notification with specified name and object.
-        var observer: NSObjectProtocol? = nil
-        func removeObserver() {
-            if let observer = observer {
-                NotificationCenter.default.removeObserver(observer)
-            }
-        }
-
-        weak var weakExpectation = expectation
-        observer = NotificationCenter
-            .default
-            .addObserver(forName: Notification.Name(rawValue: notificationName),
-                                object: objectToObserve,
-                                queue: nil,
-                                usingBlock: {
-                                    notification in
-                                    guard let expectation = weakExpectation else {
-                                        removeObserver()
-                                        return
-                                    }
-
-                                    // If the handler is invoked, the test will
-                                    // only pass if true is returned.
-                                    if let handler = handler {
-                                        if handler(notification) {
-                                            expectation.fulfill()
-                                            removeObserver()
-                                        }
-                                    } else {
-                                        expectation.fulfill()
-                                        removeObserver()
-                                    }
-            })
-
-        return expectation
-    }
-
-    /// Creates and returns an expectation that is fulfilled if the predicate
-    /// returns true when evaluated with the given object. The expectation
-    /// periodically evaluates the predicate and also may use notifications or
-    /// other events to optimistically re-evaluate.
-    ///
-    /// - Parameter predicate: The predicate that will be used to evaluate the
-    ///   object.
-    /// - Parameter object: The object that is evaluated against the conditions
-    ///   specified by the predicate.
-    /// - Parameter file: The file name to use in the error message if
-    ///   this expectation is not waited for. Default is the file
-    ///   containing the call to this method. It is rare to provide this
-    ///   parameter when calling this method.
-    /// - Parameter line: The line number to use in the error message if the
-    ///   this expectation is not waited for. Default is the line
-    ///   number of the call to this method in the calling file. It is rare to
-    ///   provide this parameter when calling this method.
-    /// - Parameter handler: A block to be invoked when evaluating the predicate
-    ///   against the object returns true. If the block is not provided the
-    ///   first successful evaluation will fulfill the expectation. If provided,
-    ///   the handler can override that behavior which leaves the caller
-    ///   responsible for fulfilling the expectation.
-    @discardableResult func expectation(for predicate: NSPredicate, evaluatedWith object: AnyObject, file: StaticString = #file, line: UInt = #line, handler: XCPredicateExpectationHandler? = nil) -> XCTestExpectation {
-        let expectation = XCPredicateExpectation(
-            predicate: predicate,
-            object: object,
-            description: "Expect `\(predicate)` for object \(object)",
-            file: file,
-            line: line,
-            testCase: self,
-            handler: handler)
-        _allExpectations.append(expectation)
-        expectation.considerFulfilling()
-        return expectation
-    }
 }
 
 internal extension XCTestCase {
